+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+
+**v1.2.0** 2020-09-11
+
+- [x] insertSome(config, select, newCommands)
+  `config = $.responsive.insertSome(config, '**/*.jpg', {jpeg: {quality: 60}});`
+- [x] webp(config)
+  `config = $.responsive.webp(config);`
+- [x] Update dependencies
+
 **v1.1.0** 2020-07-03
 
 - [x] Avoid duplicates while building config